+++ tokio-net/tests/signal_dropping_does_not_deregister_other_instances.rs
@@ -3,8 +3,8 @@
 #![warn(rust_2018_idioms)]
 #![feature(async_await)]
 
-pub mod signal_support;
-use crate::signal_support::*;
+pub mod support;
+use support::*;
 
 #[tokio::test]
 async fn dropping_signal_does_not_deregister_any_other_instances() {
