--- tokio-process/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "tokio-process"
-# When releasing to crates.io:
-# - Remove path dependencies
-# - Update html_root_url.
-# - Update doc url
-#   - Cargo.toml
-# - Update CHANGELOG.md.
-# - Create "v0.1.x" git tag.
-version = "0.3.0-alpha.1"
-edition = "2018"
-authors = ["Tokio Contributors <team@tokio.rs>"]
-license = "MIT"
-repository = "https://github.com/tokio-rs/tokio"
-homepage = "https://github.com/tokio-rs/tokio"
-documentation = "https://docs.rs/tokio-process/0.2.4/tokio_process"
-description = """
-An implementation of an asynchronous process management backed futures.
-"""
-categories = ["asynchronous"]
-
-[dependencies]
-futures-core-preview = "=0.3.0-alpha.18"
-futures-util-preview = "=0.3.0-alpha.18"
-log = "0.4"
-tokio-io = { version = "=0.2.0-alpha.2", path = "../tokio-io", features = ["util"] }
-tokio-net = { version = "=0.2.0-alpha.2", path = "../tokio-net" }
-
-[dev-dependencies.tokio]
-version = "0.2.0-alpha.1"
-path = "../tokio"
-default-features = false
-features = ["codec", "rt-full"]
-
-[target.'cfg(windows)'.dependencies]
-mio-named-pipes = "0.1"
-tokio-sync = { version = "=0.2.0-alpha.2", path = "../tokio-sync" }
-
-[target.'cfg(windows)'.dependencies.winapi]
-version = "0.3"
-features = [
-  "handleapi",
-  "winerror",
-  "minwindef",
-  "processthreadsapi",
-  "synchapi",
-  "threadpoollegacyapiset",
-  "winbase",
-  "winnt",
-]
-
-[target.'cfg(unix)'.dependencies]
-crossbeam-queue = "0.1.2"
-lazy_static = "1.3"
-libc = "0.2"
-log = "0.4"
-mio = "0.6.5"
-tokio-net = { version = "=0.2.0-alpha.2", path = "../tokio-net", features = ["signal"] }
-
-[package.metadata.docs.rs]
-all-features = true
