--- tokio-process/src/bin/exit.rs
@@ -1,5 +1,0 @@
-#[allow(dead_code)]
-
-fn main() {
-    std::process::exit(std::env::args().nth(1).unwrap().parse().unwrap());
-}
